Move `init3()` and `is_lean3_project()` back to `lean3.lua`.
